Refactor transcription reducer to delegate editor actions to transcriptionWordsReducer
