+++ src/renderer/store/transcriptionWords/opPayloads.ts
@@ -1,0 +1,17 @@
+import { IndexRange, Word } from '../../../sharedTypes';
+
+export interface DeleteSelectionPayload {
+  ranges: IndexRange[];
+}
+
+export interface PasteWordsPayload {
+  startIndex: number;
+  clipboard: Word[];
+}
+
+export type UndoDeleteSelectionPayload = DeleteSelectionPayload;
+
+export interface UndoPasteWordsPayload {
+  startIndex: number;
+  clipboardLength: number;
+}
